refactor: move From impls to vm types
